+++ cumulus/parachains/runtimes/test-utils/src/lib.rs
@@ -16,10 +16,13 @@
 use core::marker::PhantomData;
 
 use codec::{Decode, DecodeLimit};
+use cumulus_pallet_parachain_system::parachain_inherent::{
+	deconstruct_parachain_inherent_data, InboundMessagesData,
+};
 use cumulus_primitives_core::{
 	relay_chain::Slot, AbridgedHrmpChannel, ParaId, PersistedValidationData,
 };
-use cumulus_primitives_parachain_inherent::{InboundMessagesData, RawParachainInherentData};
+use cumulus_primitives_parachain_inherent::ParachainInherentData;
 use cumulus_test_relay_sproof_builder::RelayStateSproofBuilder;
 use frame_support::{
 	dispatch::{DispatchResult, GetDispatchInfo, RawOrigin},
@@ -336,7 +339,7 @@ where
 
 			let (relay_parent_storage_root, relay_chain_state) =
 				sproof_builder.into_state_root_and_proof();
-			let inherent_data = RawParachainInherentData {
+			let inherent_data = ParachainInherentData {
 				validation_data: PersistedValidationData {
 					parent_head,
 					relay_parent_number: (block_number.saturated_into::<u32>() * 2 + 1).into(),
@@ -351,7 +354,7 @@ where
 			};
 
 			let (inherent_data, downward_messages, horizontal_messages) =
-				inherent_data.deconstruct();
+				deconstruct_parachain_inherent_data(inherent_data);
 
 			let _ = cumulus_pallet_parachain_system::Pallet::<Runtime>::set_validation_data(
 				Runtime::RuntimeOrigin::none(),
@@ -712,7 +715,7 @@ pub fn mock_open_hrmp_channel<
 	// to storage; they must also be included in the inherent data.
 	let inherent_data = {
 		let mut inherent_data = InherentData::default();
-		let system_inherent_data = RawParachainInherentData {
+		let system_inherent_data = ParachainInherentData {
 			validation_data: vfp,
 			relay_chain_state,
 			downward_messages: Default::default(),
