+++ cumulus/primitives/parachain-inherent/src/lib.rs
@@ -16,30 +16,28 @@
 
 //! Cumulus parachain inherent
 //!
-//! The [`RawParachainInherentData`] is the data that is passed by the collator to the parachain
+//! The [`ParachainInherentData`] is the data that is passed by the collator to the parachain
 //! runtime. The runtime will use this data to execute messages from other parachains/the relay
 //! chain or to read data from the relay chain state. When the parachain is validated by a parachain
 //! validator on the relay chain, this data is checked for correctness. If the data passed by the
 //! collator to the runtime isn't correct, the parachain candidate is considered invalid.
 //!
-//! To create a [`RawParachainInherentData`] for a specific relay chain block, there exists the
+//! To create a [`ParachainInherentData`] for a specific relay chain block, there exists the
 //! `ParachainInherentDataExt` trait in `cumulus-client-parachain-inherent` that helps with this.
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
 extern crate alloc;
 
-use alloc::{collections::btree_map::BTreeMap, vec, vec::Vec};
-use core::fmt::Debug;
+use alloc::{collections::btree_map::BTreeMap, vec::Vec};
 use cumulus_primitives_core::{
 	relay_chain::{
 		vstaging::ApprovedPeerId, BlakeTwo256, BlockNumber as RelayChainBlockNumber,
-		Hash as RelayHash, HashT as _, Header as RelayHeader, InboundMessageId,
+		Hash as RelayHash, HashT as _, Header as RelayHeader,
 	},
 	InboundDownwardMessage, InboundHrmpMessage, ParaId, PersistedValidationData,
 };
 use scale_info::TypeInfo;
-use sp_core::{bounded::BoundedSlice, Get};
 use sp_inherents::InherentIdentifier;
 
 /// The identifier for the parachain inherent.
@@ -66,7 +64,7 @@ pub mod v0 {
 		PartialEq,
 		TypeInfo,
 	)]
-	pub struct RawParachainInherentData {
+	pub struct ParachainInherentData {
 		pub validation_data: PersistedValidationData,
 		/// A storage proof of a predefined set of keys from the relay-chain.
 		///
@@ -87,274 +85,6 @@ pub mod v0 {
 	}
 }
 
-/// An inbound message whose content was hashed.
-#[derive(
-	codec::Encode,
-	codec::Decode,
-	codec::DecodeWithMemTracking,
-	sp_core::RuntimeDebug,
-	Clone,
-	PartialEq,
-	TypeInfo,
-)]
-pub struct HashedMessage {
-	pub sent_at: RelayChainBlockNumber,
-	pub msg_hash: sp_core::H256,
-}
-
-impl From<&InboundDownwardMessage<RelayChainBlockNumber>> for HashedMessage {
-	fn from(msg: &InboundDownwardMessage<RelayChainBlockNumber>) -> Self {
-		Self { sent_at: msg.sent_at, msg_hash: MessageQueueChain::hash_msg(&msg.msg) }
-	}
-}
-
-impl From<&InboundHrmpMessage> for HashedMessage {
-	fn from(msg: &InboundHrmpMessage) -> Self {
-		Self { sent_at: msg.sent_at, msg_hash: MessageQueueChain::hash_msg(&msg.data) }
-	}
-}
-
-pub trait InboundMessage<BlockNumber> {
-	type CompressedMessage: Debug;
-
-	fn data(&self) -> &[u8];
-
-	fn sent_at(&self) -> BlockNumber;
-
-	fn to_compressed(&self) -> Self::CompressedMessage;
-}
-
-/// A collection of inbound messages.
-#[derive(
-	codec::Encode,
-	codec::Decode,
-	codec::DecodeWithMemTracking,
-	sp_core::RuntimeDebug,
-	Clone,
-	PartialEq,
-	TypeInfo,
-)]
-pub struct InboundMessagesCollection<Message: InboundMessage<RelayChainBlockNumber>> {
-	messages: Vec<Message>,
-}
-
-impl<Message: InboundMessage<RelayChainBlockNumber>> InboundMessagesCollection<Message> {
-	pub fn new(messages: Vec<Message>) -> Self {
-		Self { messages }
-	}
-
-	/// Drop all the messages up to `last_processed_msg`.
-	pub fn drop_processed_messages(
-		&mut self,
-		last_processed_msg: &InboundMessageId<RelayChainBlockNumber>,
-	) {
-		let mut last_processed_msg_idx = None;
-		let messages = &mut self.messages;
-		for (rev_idx, message) in messages.iter().rev().enumerate() {
-			let idx = (messages.len() - rev_idx - 1) as u32;
-			let sent_at = message.sent_at();
-			if sent_at == last_processed_msg.sent_at {
-				last_processed_msg_idx = idx.checked_sub(last_processed_msg.reverse_idx);
-				break;
-			}
-			if sent_at < last_processed_msg.sent_at {
-				last_processed_msg_idx = Some(idx);
-				break;
-			}
-		}
-		if let Some(last_processed_msg_idx) = last_processed_msg_idx {
-			messages.drain(..last_processed_msg_idx as usize + 1);
-		}
-	}
-
-	pub fn into_abridged(
-		self,
-		size_limit: &mut usize,
-	) -> AbridgedInboundMessagesCollection<Message> {
-		let mut messages = self.messages;
-
-		let mut maybe_split_off_pos = None;
-		for (idx, message) in messages.iter().enumerate() {
-			if *size_limit < message.data().len() {
-				break;
-			}
-			*size_limit -= message.data().len();
-
-			maybe_split_off_pos = Some(idx + 1);
-		}
-
-		let mut compressed_messages = vec![];
-		if let Some(split_off_pos) = maybe_split_off_pos {
-			let extra_messages = messages.split_off(split_off_pos);
-			compressed_messages = extra_messages.iter().map(|msg| msg.to_compressed()).collect();
-		}
-
-		AbridgedInboundMessagesCollection {
-			full_messages: messages,
-			hashed_messages: compressed_messages,
-		}
-	}
-}
-
-/// A compressed collection of inbound messages.
-///
-/// The first messages in the collection (up to a limit) contain the full message data.
-/// The messages that exceed that limit are hashed.
-#[derive(
-	codec::Encode,
-	codec::Decode,
-	codec::DecodeWithMemTracking,
-	sp_core::RuntimeDebug,
-	Clone,
-	PartialEq,
-	TypeInfo,
-)]
-pub struct AbridgedInboundMessagesCollection<Message: InboundMessage<RelayChainBlockNumber>> {
-	full_messages: Vec<Message>,
-	hashed_messages: Vec<Message::CompressedMessage>,
-}
-
-impl<Message: InboundMessage<RelayChainBlockNumber>> AbridgedInboundMessagesCollection<Message> {
-	pub fn messages(&self) -> (&[Message], &[Message::CompressedMessage]) {
-		(&self.full_messages, &self.hashed_messages)
-	}
-
-	pub fn check_advancement_rule(&self, collection_name: &str, max_size: usize) {
-		if self.hashed_messages.len() > 0 {
-			let mut size = 0usize;
-			for msg in &self.full_messages {
-				size = size.saturating_add(msg.data().len());
-			}
-			let min_size = ((max_size as f64) * 0.75) as usize;
-
-			assert!(
-				size >= min_size,
-				"[{}] Advancement rule violation: mandatory messages size smaller than expected: \
-				{} < {}",
-				collection_name,
-				size,
-				min_size
-			);
-		}
-	}
-}
-
-impl<Message: InboundMessage<RelayChainBlockNumber>> Default
-	for AbridgedInboundMessagesCollection<Message>
-{
-	fn default() -> Self {
-		Self { full_messages: vec![], hashed_messages: vec![] }
-	}
-}
-
-impl InboundMessage<RelayChainBlockNumber> for InboundDownwardMessage<RelayChainBlockNumber> {
-	type CompressedMessage = HashedMessage;
-
-	fn data(&self) -> &[u8] {
-		&self.msg
-	}
-
-	fn sent_at(&self) -> RelayChainBlockNumber {
-		self.sent_at
-	}
-
-	fn to_compressed(&self) -> Self::CompressedMessage {
-		self.into()
-	}
-}
-
-pub type InboundDownwardMessages =
-	InboundMessagesCollection<InboundDownwardMessage<RelayChainBlockNumber>>;
-
-pub type AbridgedInboundDownwardMessages =
-	AbridgedInboundMessagesCollection<InboundDownwardMessage<RelayChainBlockNumber>>;
-
-impl AbridgedInboundDownwardMessages {
-	pub fn bounded_msgs_iter<MaxMessageLen: Get<u32>>(
-		&self,
-	) -> impl Iterator<Item = BoundedSlice<u8, MaxMessageLen>> {
-		self.full_messages
-			.iter()
-			// Note: we are not using `.defensive()` here since that prints the whole value to
-			// console. In case that the message is too long, this clogs up the log quite badly.
-			.filter_map(|m| match BoundedSlice::try_from(&m.msg[..]) {
-				Ok(bounded) => Some(bounded),
-				Err(_) => {
-					log::error!("Inbound Downward message was too long; dropping");
-					debug_assert!(false);
-					None
-				},
-			})
-	}
-}
-
-impl InboundMessage<RelayChainBlockNumber> for (ParaId, InboundHrmpMessage) {
-	type CompressedMessage = (ParaId, HashedMessage);
-
-	fn data(&self) -> &[u8] {
-		&self.1.data
-	}
-
-	fn sent_at(&self) -> RelayChainBlockNumber {
-		self.1.sent_at
-	}
-
-	fn to_compressed(&self) -> Self::CompressedMessage {
-		let (sender, message) = self;
-		(*sender, message.into())
-	}
-}
-
-pub type InboundHrmpMessages = InboundMessagesCollection<(ParaId, InboundHrmpMessage)>;
-
-impl InboundHrmpMessages {
-	// Prepare horizontal messages for a more convenient processing:
-	//
-	// Instead of a mapping from a para to a list of inbound HRMP messages, we will have a
-	// list of tuples `(sender, message)` first ordered by `sent_at` (the relay chain block
-	// number in which the message hit the relay-chain) and second ordered by para id
-	// ascending.
-	pub fn from_map(messages_map: BTreeMap<ParaId, Vec<InboundHrmpMessage>>) -> Self {
-		let mut messages = messages_map
-			.into_iter()
-			.flat_map(|(sender, channel_contents)| {
-				channel_contents.into_iter().map(move |message| (sender, message))
-			})
-			.collect::<Vec<_>>();
-		messages.sort_by(|(sender_a, msg_a), (sender_b, msg_b)| {
-			// first sort by sent-at and then by the para id
-			(msg_a.sent_at, sender_a).cmp(&(msg_b.sent_at, sender_b))
-		});
-
-		Self { messages }
-	}
-}
-
-pub type AbridgedInboundHrmpMessages =
-	AbridgedInboundMessagesCollection<(ParaId, InboundHrmpMessage)>;
-
-impl AbridgedInboundHrmpMessages {
-	pub fn get_senders(&self) -> Vec<ParaId> {
-		let mut senders = vec![];
-
-		let messages = self.full_messages.iter().map(|(sender, _msg)| sender);
-		let hashed_messages = self.hashed_messages.iter().map(|(sender, _msg)| sender);
-		for sender in messages.chain(hashed_messages) {
-			if let Err(pos) = senders.binary_search(sender) {
-				senders.insert(pos, *sender);
-			}
-		}
-
-		senders
-	}
-
-	pub fn flat_msgs_iter(&self) -> impl Iterator<Item = (ParaId, RelayChainBlockNumber, &[u8])> {
-		self.full_messages
-			.iter()
-			.map(|&(sender, ref message)| (sender, message.sent_at, &message.data[..]))
-	}
-}
-
 /// The inherent data that is passed by the collator to the parachain runtime.
 #[derive(
 	codec::Encode,
@@ -365,7 +95,7 @@ impl AbridgedInboundHrmpMessages {
 	PartialEq,
 	TypeInfo,
 )]
-pub struct RawParachainInherentData {
+pub struct ParachainInherentData {
 	pub validation_data: PersistedValidationData,
 	/// A storage proof of a predefined set of keys from the relay-chain.
 	///
@@ -393,9 +123,9 @@ pub struct RawParachainInherentData {
 }
 
 // Upgrades the ParachainInherentData v0 to the newest format.
-impl Into<RawParachainInherentData> for v0::RawParachainInherentData {
-	fn into(self) -> RawParachainInherentData {
-		RawParachainInherentData {
+impl Into<ParachainInherentData> for v0::ParachainInherentData {
+	fn into(self) -> ParachainInherentData {
+		ParachainInherentData {
 			validation_data: self.validation_data,
 			relay_chain_state: self.relay_chain_state,
 			downward_messages: self.downward_messages,
@@ -407,11 +137,11 @@ impl Into<RawParachainInherentData> for v0::RawParachainInherentData {
 }
 
 #[cfg(feature = "std")]
-impl RawParachainInherentData {
-	/// Transforms [`RawParachainInherentData`] into [`v0::RawParachainInherentData`]. Can be used
+impl ParachainInherentData {
+	/// Transforms [`ParachainInherentData`] into [`v0::ParachainInherentData`]. Can be used
 	/// to create inherent data compatible with old runtimes.
-	fn as_v0(&self) -> v0::RawParachainInherentData {
-		v0::RawParachainInherentData {
+	fn as_v0(&self) -> v0::ParachainInherentData {
+		v0::ParachainInherentData {
 			validation_data: self.validation_data.clone(),
 			relay_chain_state: self.relay_chain_state.clone(),
 			downward_messages: self.downward_messages.clone(),
@@ -422,7 +152,7 @@ impl RawParachainInherentData {
 
 #[cfg(feature = "std")]
 #[async_trait::async_trait]
-impl sp_inherents::InherentDataProvider for RawParachainInherentData {
+impl sp_inherents::InherentDataProvider for ParachainInherentData {
 	async fn provide_inherent_data(
 		&self,
 		inherent_data: &mut sp_inherents::InherentData,
@@ -440,63 +170,30 @@ impl sp_inherents::InherentDataProvider for RawParachainInherentData {
 	}
 }
 
-impl RawParachainInherentData {
-	pub fn deconstruct(
-		self,
-	) -> (ParachainInherentData, InboundDownwardMessages, InboundHrmpMessages) {
-		(
-			ParachainInherentData {
-				validation_data: self.validation_data,
-				relay_chain_state: self.relay_chain_state,
-				relay_parent_descendants: self.relay_parent_descendants,
-				collator_peer_id: self.collator_peer_id,
-			},
-			InboundDownwardMessages::new(self.downward_messages),
-			InboundHrmpMessages::from_map(self.horizontal_messages),
-		)
+/// An inbound message whose content was hashed.
+#[derive(
+	codec::Encode,
+	codec::Decode,
+	codec::DecodeWithMemTracking,
+	sp_core::RuntimeDebug,
+	Clone,
+	PartialEq,
+	TypeInfo,
+)]
+pub struct HashedMessage {
+	pub sent_at: RelayChainBlockNumber,
+	pub msg_hash: sp_core::H256,
+}
+
+impl From<&InboundDownwardMessage<RelayChainBlockNumber>> for HashedMessage {
+	fn from(msg: &InboundDownwardMessage<RelayChainBlockNumber>) -> Self {
+		Self { sent_at: msg.sent_at, msg_hash: MessageQueueChain::hash_msg_data(&msg.msg) }
 	}
 }
 
-/// The basic inherent data that is passed by the collator to the parachain runtime.
-/// This data doesn't contain any messages.
-#[derive(
-	codec::Encode,
-	codec::Decode,
-	codec::DecodeWithMemTracking,
-	sp_core::RuntimeDebug,
-	Clone,
-	PartialEq,
-	TypeInfo,
-)]
-pub struct ParachainInherentData {
-	pub validation_data: PersistedValidationData,
-	pub relay_chain_state: sp_trie::StorageProof,
-	pub relay_parent_descendants: Vec<RelayHeader>,
-	pub collator_peer_id: Option<ApprovedPeerId>,
-}
-
-/// The messages that are passed by the collator to the parachain runtime as part of the
-/// inherent data.
-#[derive(
-	codec::Encode,
-	codec::Decode,
-	codec::DecodeWithMemTracking,
-	sp_core::RuntimeDebug,
-	Clone,
-	PartialEq,
-	TypeInfo,
-)]
-pub struct InboundMessagesData {
-	pub downward_messages: AbridgedInboundDownwardMessages,
-	pub horizontal_messages: AbridgedInboundHrmpMessages,
-}
-
-impl InboundMessagesData {
-	pub fn new(
-		dmq_msgs: AbridgedInboundDownwardMessages,
-		hrmp_msgs: AbridgedInboundHrmpMessages,
-	) -> Self {
-		Self { downward_messages: dmq_msgs, horizontal_messages: hrmp_msgs }
+impl From<&InboundHrmpMessage> for HashedMessage {
+	fn from(msg: &InboundHrmpMessage) -> Self {
+		Self { sent_at: msg.sent_at, msg_hash: MessageQueueChain::hash_msg_data(&msg.data) }
 	}
 }
 
@@ -521,7 +218,8 @@ impl MessageQueueChain {
 		Self(hash)
 	}
 
-	fn hash_msg(msg: &Vec<u8>) -> sp_core::H256 {
+	/// Hash the provided message data.
+	fn hash_msg_data(msg: &Vec<u8>) -> sp_core::H256 {
 		BlakeTwo256::hash_of(msg)
 	}
 
@@ -548,142 +246,5 @@ impl MessageQueueChain {
 	/// This is agreed to be the zero hash for an empty chain.
 	pub fn head(&self) -> RelayHash {
 		self.0
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	fn build_inbound_dm_vec(
-		info: &[(RelayChainBlockNumber, usize)],
-	) -> Vec<InboundDownwardMessage<RelayChainBlockNumber>> {
-		let mut messages = vec![];
-		for (sent_at, size) in info.iter() {
-			let data = vec![1; *size];
-			messages.push(InboundDownwardMessage { sent_at: *sent_at, msg: data })
-		}
-		messages
-	}
-
-	#[test]
-	fn drop_processed_messages_works() {
-		let msgs_vec =
-			build_inbound_dm_vec(&[(0, 0), (0, 0), (2, 0), (2, 0), (2, 0), (2, 0), (3, 0)]);
-
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 3, reverse_idx: 0 });
-		assert_eq!(msgs.messages, []);
-
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 2, reverse_idx: 0 });
-		assert_eq!(msgs.messages, msgs_vec[6..]);
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 2, reverse_idx: 1 });
-		assert_eq!(msgs.messages, msgs_vec[5..]);
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 2, reverse_idx: 4 });
-		assert_eq!(msgs.messages, msgs_vec[2..]);
-
-		// Go back starting from the last message sent at block 2, with 1 more message than the
-		// total number of messages sent at 2.
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 2, reverse_idx: 5 });
-		assert_eq!(msgs.messages, msgs_vec[1..]);
-
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 0, reverse_idx: 1 });
-		assert_eq!(msgs.messages, msgs_vec[1..]);
-		// Go back starting from the last message sent at block 0, with 1 more message than the
-		// total number of messages sent at 0.
-		let mut msgs = InboundDownwardMessages::new(msgs_vec.clone());
-		msgs.drop_processed_messages(&InboundMessageId { sent_at: 0, reverse_idx: 3 });
-		assert_eq!(msgs.messages, msgs_vec);
-	}
-
-	#[test]
-	fn into_compressed_works() {
-		let msgs_vec = build_inbound_dm_vec(&[(0, 100), (0, 100), (0, 150), (0, 50)]);
-		let msgs = InboundDownwardMessages::new(msgs_vec.clone());
-
-		let mut size_limit = 150;
-		let compressed_msgs = msgs.clone().into_abridged(&mut size_limit);
-		assert_eq!(size_limit, 50);
-		assert_eq!(&compressed_msgs.full_messages, &msgs_vec[..1]);
-		assert_eq!(
-			compressed_msgs.hashed_messages,
-			vec![(&msgs_vec[1]).into(), (&msgs_vec[2]).into(), (&msgs_vec[3]).into()]
-		);
-
-		let mut size_limit = 200;
-		let compressed_msgs = msgs.clone().into_abridged(&mut size_limit);
-		assert_eq!(size_limit, 0);
-		assert_eq!(&compressed_msgs.full_messages, &msgs_vec[..2]);
-		assert_eq!(
-			compressed_msgs.hashed_messages,
-			vec![(&msgs_vec[2]).into(), (&msgs_vec[3]).into()]
-		);
-
-		let mut size_limit = 399;
-		let compressed_msgs = msgs.clone().into_abridged(&mut size_limit);
-		assert_eq!(size_limit, 49);
-		assert_eq!(&compressed_msgs.full_messages, &msgs_vec[..3]);
-		assert_eq!(compressed_msgs.hashed_messages, vec![(&msgs_vec[3]).into()]);
-
-		let mut size_limit = 400;
-		let compressed_msgs = msgs.clone().into_abridged(&mut size_limit);
-		assert_eq!(size_limit, 0);
-		assert_eq!(&compressed_msgs.full_messages, &msgs_vec);
-		assert_eq!(compressed_msgs.hashed_messages, vec![]);
-	}
-
-	#[test]
-	fn from_map_works() {
-		let mut messages_map: BTreeMap<ParaId, Vec<InboundHrmpMessage>> = BTreeMap::new();
-		messages_map.insert(
-			1000.into(),
-			vec![
-				InboundHrmpMessage { sent_at: 0, data: vec![0] },
-				InboundHrmpMessage { sent_at: 0, data: vec![1] },
-				InboundHrmpMessage { sent_at: 1, data: vec![2] },
-			],
-		);
-		messages_map.insert(
-			2000.into(),
-			vec![
-				InboundHrmpMessage { sent_at: 0, data: vec![3] },
-				InboundHrmpMessage { sent_at: 0, data: vec![4] },
-				InboundHrmpMessage { sent_at: 1, data: vec![5] },
-			],
-		);
-		messages_map.insert(
-			3000.into(),
-			vec![
-				InboundHrmpMessage { sent_at: 0, data: vec![6] },
-				InboundHrmpMessage { sent_at: 1, data: vec![7] },
-				InboundHrmpMessage { sent_at: 2, data: vec![8] },
-				InboundHrmpMessage { sent_at: 3, data: vec![9] },
-				InboundHrmpMessage { sent_at: 4, data: vec![10] },
-			],
-		);
-
-		let msgs = InboundHrmpMessages::from_map(messages_map);
-		assert_eq!(
-			msgs.messages,
-			[
-				(1000.into(), InboundHrmpMessage { sent_at: 0, data: vec![0] }),
-				(1000.into(), InboundHrmpMessage { sent_at: 0, data: vec![1] }),
-				(2000.into(), InboundHrmpMessage { sent_at: 0, data: vec![3] }),
-				(2000.into(), InboundHrmpMessage { sent_at: 0, data: vec![4] }),
-				(3000.into(), InboundHrmpMessage { sent_at: 0, data: vec![6] }),
-				(1000.into(), InboundHrmpMessage { sent_at: 1, data: vec![2] }),
-				(2000.into(), InboundHrmpMessage { sent_at: 1, data: vec![5] }),
-				(3000.into(), InboundHrmpMessage { sent_at: 1, data: vec![7] }),
-				(3000.into(), InboundHrmpMessage { sent_at: 2, data: vec![8] }),
-				(3000.into(), InboundHrmpMessage { sent_at: 3, data: vec![9] }),
-				(3000.into(), InboundHrmpMessage { sent_at: 4, data: vec![10] })
-			]
-		)
 	}
 }
