+++ cumulus/test/service/src/bench_utils.rs
@@ -19,7 +19,9 @@ use codec::Encode;
 use sc_block_builder::BlockBuilderBuilder;
 
 use crate::{construct_extrinsic, Client as TestClient};
-use cumulus_client_parachain_inherent::{InboundMessagesData, ParachainInherentData};
+use cumulus_pallet_parachain_system::parachain_inherent::{
+	BasicParachainInherentData, InboundMessagesData,
+};
 use cumulus_primitives_core::{relay_chain::AccountId, PersistedValidationData};
 use cumulus_test_relay_sproof_builder::RelayStateSproofBuilder;
 use cumulus_test_runtime::{
@@ -88,7 +90,7 @@ pub fn extrinsic_set_validation_data(
 	};
 
 	let (relay_parent_storage_root, relay_chain_state) = sproof_builder.into_state_root_and_proof();
-	let data = ParachainInherentData {
+	let data = BasicParachainInherentData {
 		validation_data: PersistedValidationData {
 			parent_head,
 			relay_parent_number: 10,
