+++ cumulus/xcm/xcm-emulator/src/lib.rs
@@ -64,13 +64,14 @@ pub use sp_tracing;
 
 // Cumulus
 pub use cumulus_pallet_parachain_system::{
+	parachain_inherent::{deconstruct_parachain_inherent_data, InboundMessagesData},
 	Call as ParachainSystemCall, Pallet as ParachainSystemPallet,
 };
 pub use cumulus_primitives_core::{
 	relay_chain::{BlockNumber as RelayBlockNumber, HeadData, HrmpChannelId},
 	AbridgedHrmpChannel, DmpMessageHandler, ParaId, PersistedValidationData, XcmpMessageHandler,
 };
-pub use cumulus_primitives_parachain_inherent::{InboundMessagesData, RawParachainInherentData};
+pub use cumulus_primitives_parachain_inherent::ParachainInherentData;
 pub use cumulus_test_relay_sproof_builder::RelayStateSproofBuilder;
 pub use pallet_message_queue::{Config as MessageQueueConfig, Pallet as MessageQueuePallet};
 pub use parachains_common::{AccountId, Balance};
@@ -200,7 +201,7 @@ pub trait Network {
 		para_id: u32,
 		relay_parent_number: u32,
 		parent_head_data: HeadData,
-	) -> RawParachainInherentData;
+	) -> ParachainInherentData;
 	fn send_horizontal_messages<I: Iterator<Item = (ParaId, RelayBlockNumber, Vec<u8>)>>(
 		to_para_id: u32,
 		iter: I,
@@ -712,7 +713,8 @@ macro_rules! decl_test_parachains {
 
 						// 1. inherent: cumulus_pallet_parachain_system::Call::set_validation_data
 						let data = N::hrmp_channel_parachain_inherent_data(para_id, relay_block_number, parent_head_data);
-						let (data, mut downward_messages, mut horizontal_messages) = data.deconstruct();
+						let (data, mut downward_messages, mut horizontal_messages) =
+							$crate::deconstruct_parachain_inherent_data(data);
 						let inbound_messages_data = $crate::InboundMessagesData::new(
 							downward_messages.into_abridged(&mut usize::MAX.clone()),
 							horizontal_messages.into_abridged(&mut usize::MAX.clone()),
@@ -1181,7 +1183,7 @@ macro_rules! decl_test_networks {
 					para_id: u32,
 					relay_parent_number: u32,
 					parent_head_data: $crate::HeadData,
-				) -> $crate::RawParachainInherentData {
+				) -> $crate::ParachainInherentData {
 					let mut sproof = $crate::RelayStateSproofBuilder::default();
 					sproof.para_id = para_id.into();
 					sproof.current_slot = $crate::polkadot_primitives::Slot::from(relay_parent_number as u64);
@@ -1214,7 +1216,7 @@ macro_rules! decl_test_networks {
 
 					let (relay_storage_root, proof) = sproof.into_state_root_and_proof();
 
-					$crate::RawParachainInherentData {
+					$crate::ParachainInherentData {
 						validation_data: $crate::PersistedValidationData {
 							parent_head: parent_head_data.clone(),
 							relay_parent_number,
